Bug #51508 UMC: Fix stopping UMC-Server during startup
Traceback (most recent call last):
  File "/usr/sbin/univention-management-console-server", line 248, in <module>
    umc_daemon.do_action()
  File "/usr/lib/python3/dist-packages/daemon/runner.py", line 283, in do_action
    func(self)
  File "/usr/lib/python3/dist-packages/daemon/runner.py", line 202, in _start
    self.app.run()
  File "/usr/sbin/univention-management-console-server", line 211, in run
    notifier.init(notifier.GENERIC)
  File "/usr/lib/python3/dist-packages/notifier/__init__.py", line 63, in init
    from . import nf_generic as nf_impl
  File "/usr/lib/python3/dist-packages/notifier/nf_generic.py", line 34, in <module>
    from . import dispatch
  File "<frozen importlib._bootstrap>", line 983, in _find_and_load
  File "<frozen importlib._bootstrap>", line 967, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 677, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 724, in exec_module
  File "<frozen importlib._bootstrap_external>", line 828, in get_code
  File "/usr/sbin/univention-management-console-server", line 236, in signal_terminate
    self.server.exit()
AttributeError: 'UMC_Daemon' object has no attribute 'server'

* make sure that kill always cause the process to end.

MAIN        ( ERROR   ) : FATAL ERROR: Traceback (most recent call last):
  File "/usr/lib/python2.7/dist-packages/univention/management/console/protocol/server.py", line 201, in _response
    ret = state.socket.send(data)
  File "/usr/lib/python2.7/dist-packages/OpenSSL/SSL.py", line 1255, in send
    result = _lib.SSL_write(self._ssl, buf, len(buf))
  File "/usr/sbin/univention-management-console-server", line 237, in signal_terminate
    raise SystemExit('Shutting down UMC server')
SystemExit: Shutting down UMC

* check if UNIX socket in module processes are already created:

Traceback (most recent call last):
  File "/usr/sbin/univention-management-console-module", line 121, in <module>
    notifier.loop()
  File "/usr/lib/pymodules/python2.7/notifier/nf_generic.py", line 287, in loop
    step()
  File "/usr/lib/pymodules/python2.7/notifier/nf_generic.py", line 243, in step
    if not timer[ CALLBACK ]():
  File "/usr/lib/pymodules/python2.7/univention/management/console/protocol/modserver.py", line 140, in _timer
    self._timed_out()
  File "/usr/lib/pymodules/python2.7/univention/management/console/protocol/modserver.py", line 150, in _timed_out
    self.exit()
  File "/usr/lib/pymodules/python2.7/univention/management/console/protocol/server.py", line 392, in exit
    os.unlink(self.__unix)
OSError: [Errno 2] Datei oder Verzeichnis nicht gefunden: '/var/run/univention-management-console/1621-1570639160845.socket'

* fix reloading

MAIN        ( PROCESS ) : Reloading configuration ...
MAIN        ( ERROR   ) : Traceback (most recent call last):
  File "/usr/sbin/univention-management-console-server", line 252, in <module>
    umc_daemon.do_action()
  File "/usr/lib/python3/dist-packages/daemon/runner.py", line 283, in do_action
    func(self)
  File "/usr/lib/python3/dist-packages/daemon/runner.py", line 202, in _start
    self.app.run()
  File "/usr/sbin/univention-management-console-server", line 208, in run
    from univention.management.console.protocol.server import Server
  File "/usr/lib/python3/dist-packages/univention/management/console/protocol/__init__.py", line 212, in <module>
    from .session import *  # noqa: F403,F401
  File "/usr/lib/python3/dist-packages/univention/management/console/protocol/session.py", line 65, in <module>
    from ..acl import LDAP_ACLs, ACLs
  File "/usr/lib/python3/dist-packages/univention/management/console/acl.py", line 81, in <module>
    import univention.admin.handlers.computers.domaincontroller_master as dc_master
  File "/usr/lib/python3/dist-packages/univention/admin/handlers/__init__.py", line 76, in <module>
    import univention.lib.admember
  File "/usr/lib/python3/dist-packages/univention/lib/admember.py", line 54, in <module>
    from samba.net import Net
  File "/usr/sbin/univention-management-console-server", line 234, in signal_user1
    self.server.reload()
AttributeError: 'NoneType' object has no attribute 'reload'

+++ management/univention-management-console/src/univention/management/console/protocol/server.py
@@ -220,7 +220,7 @@ class MagicBucket(object):
 		except (SSL.SysCallError, SSL.Error, socket.error) as error:
 			CRYPT.warn('SSL error in _response: %s. Probably the socket was closed by the client.' % str(error))
 			self._cleanup(state.socket)
-		except BaseException:  # close the connection to the client. we can't do anything else
+		except Exception:  # close the connection to the client. we can't do anything else
 			CORE.error('FATAL ERROR: %s' % (traceback.format_exc(),))
 			self._cleanup(state.socket)
 
@@ -274,7 +274,7 @@ class Server(signals.Provider):
 			CORE.info('Using a TCP socket')
 			try:
 				self.__realsocket = socket.socket(socket.AF_INET6, socket.SOCK_STREAM)
-			except BaseException:
+			except Exception:
 				CORE.warn('Cannot open socket with AF_INET6 (Python reports socket.has_ipv6 is %s), trying AF_INET' % socket.has_ipv6)
 				self.__realsocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 
@@ -389,7 +389,8 @@ class Server(signals.Provider):
 			self.__realsocket.close()
 			self.__realsocket = None
 		if self.__unix:
-			os.unlink(self.__unix)
+			if os.path.exists(self.__unix):
+				os.unlink(self.__unix)
 			self.__unix = None
 
 		self.__bucket = None
